hwmon: (asus_wmi_ec_sensors) read sensors as signed ints
Temperature sensor readings are signed, which is hinted by their blank
value (oxd8, 216 as unsigned and -40 as signed). T_Sensor, Crosshair
VIII Hero, and a freezer were used to confirm that.

Here we read fan sensors as signed too, because with their typical
values and 2-byte width, I can't tell a difference between signed and
unsigned, as I don't have a high speed chipset fan.

Co-developed-by: Eugene Shalygin <eugene.shalygin@gmail.com>

+++ drivers/hwmon/asus_wmi_ec_sensors.c
@@ -344,11 +344,11 @@ static inline long get_sensor_value(const struct ec_sensor_info *si, u8 *data)
 {
 	switch (si->addr.size) {
 	case 1:
-		return *data;
+		return (s8)*data;
 	case 2:
-		return get_unaligned_be16(data);
+		return (s16)get_unaligned_be16(data);
 	case 4:
-		return get_unaligned_be32(data);
+		return (s32)get_unaligned_be32(data);
 	default:
 		return 0;
 	}
